# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Simula Native Ad Server",
        "description": "Campaign/ad-set management, session resolution and native ad serving with geo + OS targeting, bandit ranking and LLM ad copy.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/adsets": {
            "post": {
                "tags": [
                    "adsets"
                ],
                "summary": "Create Ad Set",
                "operationId": "create_ad_set_adsets_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdSetCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdSetCreateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "adsets"
                ],
                "summary": "List Ad Sets",
                "operationId": "list_ad_sets_adsets_get",
                "parameters": [
                    {
                        "name": "campaign_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Campaign Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdSet"
                                    },
                                    "title": "Response List Ad Sets Adsets Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/adsets/{ad_set_id}": {
            "get": {
                "tags": [
                    "adsets"
                ],
                "summary": "Get Ad Set",
                "operationId": "get_ad_set_adsets__ad_set_id__get",
                "parameters": [
                    {
                        "name": "ad_set_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ad Set Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdSet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/adsets/{ad_set_id}/variants": {
            "get": {
                "tags": [
                    "adsets"
                ],
                "summary": "List Variants",
                "operationId": "list_variants_adsets__ad_set_id__variants_get",
                "parameters": [
                    {
                        "name": "ad_set_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ad Set Id"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Active"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdVariant"
                                    },
                                    "title": "Response List Variants Adsets  Ad Set Id  Variants Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdSet": {
                "properties": {
                    "ad_set_id": {
                        "type": "string",
                        "title": "Ad Set Id"
                    },
                    "campaign_id": {
                        "type": "string",
                        "title": "Campaign Id"
                    },
                    "ad_set_name": {
                        "type": "string",
                        "title": "Ad Set Name"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "default": true
                    },
                    "character_names": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Character Names"
                    },
                    "video_urls": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Video Urls"
                    },
                    "ctas": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Ctas"
                    },
                    "ai_prompts": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Ai Prompts"
                    },
                    "fallback_copy": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Fallback Copy"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "campaign_id",
                    "ad_set_name"
                ],
                "title": "AdSet"
            },
            "AdSetCreate": {
                "properties": {
                    "campaign_id": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Campaign Id"
                    },
                    "ad_set_name": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 1,
                        "title": "Ad Set Name"
                    },
                    "character_names": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Character Names"
                    },
                    "video_urls": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Video Urls"
                    },
                    "ctas": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Ctas"
                    },
                    "ai_prompts": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Ai Prompts"
                    },
                    "fallback_copy": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Fallback Copy"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "default": true
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "campaign_id",
                    "ad_set_name",
                    "character_names",
                    "video_urls",
                    "ctas",
                    "ai_prompts"
                ],
                "title": "AdSetCreate"
            },
            "AdSetCreateResponse": {
                "properties": {
                    "ad_set": {
                        "$ref": "#/components/schemas/AdSet"
                    },
                    "variants": {
                        "items": {
                            "$ref": "#/components/schemas/AdVariant"
                        },
                        "type": "array",
                        "title": "Variants"
                    },
                    "variant_count": {
                        "type": "integer",
                        "title": "Variant Count"
                    },
                    "campaign_linked": {
                        "type": "boolean",
                        "title": "Campaign Linked"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "ad_set",
                    "variants",
                    "variant_count",
                    "campaign_linked"
                ],
                "title": "AdSetCreateResponse"
            },
            "AdVariant": {
                "properties": {
                    "variant_id": {
                        "type": "string",
                        "title": "Variant Id"
                    },
                    "ad_set_id": {
                        "type": "string",
                        "title": "Ad Set Id"
                    },
                    "campaign_id": {
                        "type": "string",
                        "title": "Campaign Id"
                    },
                    "character_name": {
                        "type": "string",
                        "title": "Character Name"
                    },
                    "video_url": {
                        "type": "string",
                        "title": "Video Url"
                    },
                    "cta": {
                        "type": "string",
                        "title": "Cta"
                    },
                    "ai_prompt": {
                        "type": "string",
                        "title": "Ai Prompt"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "default": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "ad_set_id",
                    "campaign_id",
                    "character_name",
                    "video_url",
                    "cta",
                    "ai_prompt"
                ],
                "title": "AdVariant"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}